Proyecto-API/requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bandit==1.7.9
bcrypt==5.0.0
boltons==21.0.0
boolean.py==5.0
bracex==2.6
CacheControl==0.14.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
click-option-group==0.5.9
colorama==0.4.6
cryptography==46.0.3
cyclonedx-python-lib==7.6.2
defusedxml==0.7.1
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.2.2
face==24.0.0
fastapi==0.120.0
filelock==3.20.0
glom==22.1.0
googleapis-common-protos==1.71.0
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.27.0
idna==3.11
importlib_metadata==7.1.0
iniconfig==2.3.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
license-expression==30.4.4
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.2
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packageurl-python==0.17.5
packaging==25.0
passlib==1.7.4
peewee==3.18.2
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.7.3
pluggy==1.6.0
protobuf==4.25.8
psycopg2-binary==2.9.11
py-serializable==1.1.2
pyasn1==0.4.8
pycparser==2.23
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.2.0
python-dotenv==1.1.1
python-jose==3.4.0
python-multipart==0.0.18
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rich==13.5.3
rpds-py==0.28.0
rsa==4.9.1
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.14
semgrep==1.89.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.44
sqlmodel==0.0.21
starlette==0.48.0
stevedore==5.5.0
toml==0.10.2
tomli==2.0.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.30.5
uvloop==0.22.1
watchfiles==1.1.1
wcmatch==8.5.2
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
zipp==3.23.0
paho-mqtt==1.6.1